Chat gpt help

+++ src/main/java/com/club/padel/repository/RolRepository.java
@@ -4,5 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.club.padel.model.Rol;
 
-public interface RolRepository extends JpaRepository<Rol, Integer> {
+public interface RolRepository extends JpaRepository<Rol, Long> {
+    // Métodos de consulta personalizados si es necesario
 }
+
